+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [0.22.3] - 2016-07-17
+## Fixed
+- Virtualenv now ues any available python 3
+- move_head_absolute not passing rate to to_printer queue
+- Tests updated for new rate parameter
+
 ## [0.22.2] - 2016-07-17
 ## Fixed
 - Debian package version